+++ flex/distributed/server.py
@@ -184,7 +184,7 @@ class Server:
 
     def send_weights(self, weights: List[np.ndarray], node_ids: Optional[any] = None):
         for w in weights:
-            w = np.array(w, dtype=np.float32)
+            w = np.array(w)
 
         self._manager.broadcast(
             ServerMessage(
